Create Week 1 Code; BFS, DFS, Kosaraju
A collection of functions of two types. The first type are those that deal with reading in graphs and putting them into a form (e.g., adjacency list representation) suitable for input to other functions. The second type are functions that implement breadth-first search, depth-first search, or Kosaraju's algorithm for computing the strongly connected components of a graph.